chore: clear codes
