+++ src/lib/form.ts
@@ -1,4 +1,5 @@
 import { z } from "zod";
+import yaml from "js-yaml";
 
 export const emailSchema = z.string({ required_error: "请输入邮箱" }).email({ message: "请输入正确的邮箱" });
 export const phoneSchema = z
@@ -37,4 +38,24 @@ export function objectToFormData(obj: Record<string, any>) {
     formData.append(key, value);
   }
   return formData;
+}
+
+export function objectToYaml(obj: { [key: string]: number | string }) {
+  try {
+    if (Object.keys(obj).length < 1) return;
+    const newObj: { [key: string]: number | string | string[] } = {};
+
+    const keys = Object.keys(obj);
+    for (let i = 0; i < keys.length; i++) {
+      newObj[i + 1] =
+        typeof obj[keys[i]] == "number" ? String.fromCharCode(65 + (obj[keys[i]] as number)) : (obj[keys[i]] as string);
+    }
+    // newObj[2] = ["A", "B", "C"];
+    const yamlData = yaml.dump(newObj);
+
+    return yamlData;
+  } catch (e) {
+    console.error("Error converting object to YAML:", e);
+    return null;
+  }
 }
