--- src/lib/forward-auth-client.ts
@@ -1,13 +1,0 @@
-export const forwardAuthClient = () => {
-  const sid = document.cookie
-    .split("; ")
-    .find((row) => row.startsWith("sid="))
-    ?.split("=")[1];
-
-  return {
-    credentials: "include" as RequestCredentials,
-    headers: {
-      Authorization: `Bearer ${sid ?? ""}`,
-    },
-  };
-};
